+++ src/main/java/racingcar/view/InputView.java
@@ -3,7 +3,6 @@ package racingcar.view;
 import camp.nextstep.edu.missionutils.Console;
 import java.util.Arrays;
 import java.util.List;
-import racingcar.ErrorMessage;
 
 public class InputView {
     private static final String DELIMITER = ",";
