+++ src/test/java/racingcar/model/RaceCarsTest.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.ErrorMessage;
+import racingcar.view.ErrorMessage;
 
 class RaceCarsTest {
     @Test
