+++ src/test/java/racingcar/model/RoundCountTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.ErrorMessage;
+import racingcar.view.ErrorMessage;
 
 class RoundCountTest {
     @Test
